Move `FieldArrayMeta` into `_fields/_meta.py`
