+++ galois/_domains/_meta.py
@@ -1,5 +1,5 @@
 """
-A module that defines the abstract base class Array.
+A module that defines the metaclass for the abstract base class Array.
 """
 from __future__ import annotations
